Fixed drone flight to work with deltaTime
- Also expanded arena to facilitate testing
- Made minor changes in Render.cpp to show debug circles

+++ Drone.h
@@ -14,16 +14,16 @@ public:
 	virtual void run(float DeltaTime);
 
 	float turnForce;
-	int chaseSpeed;
-	int controlSpeed;
-	int controlRange;		//how far out the drone will get before beginning to chase again
-	int matchRange;
+	float chaseSpeed;
+	float controlSpeed;
+	float controlRange;		//how far out the drone will get before beginning to chase again
+	float matchRange;
 	float speedSmoother;	//moves between chaseSpeed (full speed) and controlSpeed (about the velocity of the ship you're following)
 	Ship* flyAround;
 	Vec2D circle;
 	float theta;
 	Vec2D* flightTarget;	//I want the drone to either chase the ship, or fly around it, so this will be set to the target ship's loc or circle
-	unsigned int ROF;
+	float ROF;
 
 };
